Add files via upload
The Amazon clone project is a front-end web development task that aims to replicate the user interface of the Amazon website. This involves designing various pages such as the homepage, product listing page, product detail page, shopping cart, and checkout page. The project focuses on using HTML for the structure and CSS for the styling.